CR fixes, comments, refactoring
